[Camera page]Step 1: Added half of the camera functions in a separate file and imported them. Status: Everything works!
