+++ br-webui/index.js
@@ -47,102 +47,35 @@ app.use(expressLiquid.middleware);
 var _companion_directory = process.env.COMPANION_DIR;
 
 var v4l2camera = require("v4l2camera");
-// This holds all of the cameras/settings detected at start, and that are currently in use, we need to update this every time we modify the camera setttings
-var _cameras = []
 
+var cameranew = require("./camera.js");
+
+cameranew.currentformat();
+
+// Load saved user camera/streaming profiles
+cameranew.loadprofiles();
+
+// Load the last known camera settings
+cameranew.loadlastsettings();
+
+// Create camera objects, set all camera settings on all cameras to the
+// last known settings
+cameranew.createnewobject();
+
+// This holds all of the cameras/settings detected at start, and that are currently in use, we need to update this every time we modify the camera setttings
+//var _cameras = cameranew._cameras
 
 //This holds the current frame size, frame rate, video device, and video format
 //These settings are passed to the video streaming application, and are used by
 //gstreamer v4l2src element. The v4l2src element needs to call the appropriate ioctls,
 //so we don't do that in this application.
-var _activeFormat
+//var _activeFormat = cameranew._activeFormat
 
-try {
-	var file_path = home_dir+"/vidformat.param";
-	var file_data = fs.readFileSync(file_path).toString();
-	var fields = file_data.split("\n");
-	_activeFormat = { "frameSize": fields[0] + "x" + fields[1], "frameRate": fields[2], "device": fields[3], "format": "H264" }
-} catch (err) {
-	logger.log("error loading video format from file", err);
-}
-
-// This holds the user created camera/streaming profiles
-var _profiles = {};
-
-// Load saved user camera/streaming profiles
-try {
-	var file_path = home_dir+"/camera-profiles";
-	_profiles = JSON.parse(fs.readFileSync(file_path).toString());
-	logger.log("loading profiles from file", _profiles);
-} catch (err) {
-	logger.log("error loading video profiles from file", err);
-}
+//This holds the user created camera/streaming profiles
+//var _profiles = cameranew._profiles;
 
 //This holds all of the last used/known settings from previous run
-var old_cameras = []
-const camera_settings_path = home_dir+"/camera-settings"
-// Load the last known camera settings
-try {
-	var file_data = fs.readFileSync(camera_settings_path);
-	old_cameras = JSON.parse(file_data.toString());
-} catch (err) {
-	logger.log("error loading file", err);
-}
-
-// Create camera objects, set all camera settings on all cameras to the
-// last known settings
-for (var i = 0; ;i++) {
-	try {
-		var cam = new v4l2camera.Camera("/dev/video" + i)
-		logger.log("found camera:", i);
-		
-		// TODO put this in driver layer
-		cam.controls.forEach(function(control) {
-			if (control.type != "class") {
-				control.value = cam.controlGet(control.id); // store all the current values locally so that we can update the frontend
-				logger.log("getting control:", control.name, control.type, control.value);
-				// HACK, some v4l2 devices report bogus default values that are way beyond
-				// min/max range, so we need to record what the default value actually is
-				// The cameras that I have seen 
-				control.default = control.value;
-				
-				// HACK, bogus max on rpi cam
-				if (control.name == "H264 I-Frame Period" && control.max > 2000000) {
-					control.max = 120;
-				}
-			}
-		});
-		
-		var has_h264 = false;
-		cam.formats.forEach(function(format) {
-			if (format.formatName == "H264") {
-				has_h264 = true;
-			}
-		});
-		
-		if (has_h264) {
-			_cameras.push(cam);
-		}
-	} catch(err) { // this is thrown once /dev/video<i> does not exist, we have enumerated all of the cameras
-		old_cameras.forEach(function(oldcam) { // Configure cameras to match last known/used settings
-			_cameras.forEach(function(cam) {
-				if (cam.device == oldcam.device) {
-					logger.log("oldcam match:", oldcam.device);
-					oldcam.controls.forEach(function(control) {
-						logger.log("setting control:", control.name, control.value);
-						try {
-							cam.controlSet(control.id, control.value);
-						} catch (err) {
-							logger.log("control set failed");
-						}
-					});
-				}
-			});
-		});
-		break;
-	}
-}
-
+//var old_cameras = cameranew.old_cameras;
 
 ////////////////// Routes
